+++ ophidian/entity_system/CMakeLists.txt
@@ -1,4 +1,7 @@
-aux_source_directory(. SRC_LIST)
-add_library(ophidian_entitysystem ${SRC_LIST})
+file(GLOB ophidian_entitysystem_SRC
+    "*.h"
+    "*.cpp"
+)
+add_library(ophidian_entitysystem ${ophidian_entitysystem_SRC})
 install(TARGETS ophidian_entitysystem DESTINATION lib)
 install(FILES EntitySystem.h Property.h DESTINATION include/ophidian/entity_system)
