--- ophidian/entity_system/Composition.cpp
@@ -1,10 +1,0 @@
-#include "Composition.h"
-
-namespace ophidian
-{
-    namespace entity_system
-    {
-
-    }
-
-}
