--- ophidian/entity_system/Property.cpp
@@ -1,7 +1,0 @@
-#include "Property.h"
-
-namespace ophidian {
-namespace entity_system {
-
-}
-}
